requirements-test.txt
# Testing dependencies for Model Compass
pytest>=7.0.0
pytest-cov>=4.0.0
pyyaml>=6.0